fix: 버그 및 타입 수정
