--- src/components/common/UserCard/style.ts
@@ -1,27 +1,0 @@
-import styled from '@emotion/styled';
-
-export const CardWrapper = styled.div`
-  position: relative;
-  display: flex;
-  align-items: center;
-  border: 1px solid black;
-  padding: 20px;
-`;
-
-export const Text = styled.div`
-  padding-left: 20px;
-  flex-grow: 1;
-  font-size: 64px;
-`;
-
-export const MbtiTag = styled.div`
-  background-color: lightgray;
-  font-size: 56px;
-  padding: 10px;
-  border-radius: 25%;
-`;
-
-export const UserCardGroupWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
