+++ src/components/search/UserCard/index.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
+import { useNavigate } from 'react-router-dom';
 import PropTypes from 'prop-types';
-import Image, { modeType } from '../Image';
+import Image, { modeType } from '~/components/common/Image';
 import UserCardGroup from './UserCardGroup';
 import { CardWrapper, MbtiTag, Text } from './style';
 
@@ -10,6 +11,7 @@ export interface UserCardProp {
   imageSize?: number;
   nickname: string;
   mbti: string;
+  id: string;
   style?: React.CSSProperties;
 }
 
@@ -19,15 +21,22 @@ const UserCard = ({
   imageSize = 200,
   nickname,
   mbti,
+  id,
   ...props
 }: UserCardProp) => {
+  const navigate = useNavigate();
+
   const imageStyle = {
     width: imageSize,
     height: imageSize,
     mode: 'cover' as modeType,
   };
+
+  const handleClick = () => {
+    navigate(`/user/${id}`);
+  };
   return (
-    <CardWrapper {...props}>
+    <CardWrapper {...props} onClick={handleClick}>
       <Image {...imageStyle} src={src} alt={alt} shape="circle" />
       <Text>{nickname}</Text>
       <MbtiTag>{mbti}</MbtiTag>
