+++ src/pages/SearchPage/index.tsx
@@ -8,7 +8,7 @@ const SearchWrapper = styled.div`
   display: flex;
   flex-direction: column;
   width: 100%;
-  height: 100vh;
+  height: calc(100vh - 90px);
   overflow: hidden;
 `;
